Remove duplicate dependencies
Problem: Two dependencies are listed in both `dependencies` and
`devDependencies`, which invokes and npm warning and can cause problems
if the required versions aren't exactly the same.

Solution: Remove the dependencies from `devDependencies`, run
`npm install`, and `npm dedupe` to try to minimize the size of the new
dependency graph.