Add changes necessary to get CSV import errors
Problem: There `savePerson()` command has been renamed to
`saveProfile()`, so we had lots of errors around how we pull the
property out of that object.

Solution: Rename references to `savePerson` to `saveProfile` around the
code to try to resolve these errors, also throw an error ('BAD') when we
get a GraphQL error from the back-end.