Add CSV tests and upload profiles + links slowly
Problem: There are three bugs that we need to circumvent: SSB-Tribes
misbehaves when you publish quickly, SSB-DB misbehaves when you publish
quickly, and SSB-GraphQL/whakapapa duplicates entries when you call
get(). This **does not** contain upgraded version of SSB-DB or
SSB-GraphQL/whakapapa, both of which are necessary.

Solution: Publish *very* slowly. While debugging I've written some tests
for the CSV import and aren't harmful to add, but the important bit is
that the app uploads records very slowly.