Link in dependencies to fix CSV import
Problem: I was using lots of `npm link` magic to get my previous commit
working.

Solution: I've put the URLs in `package.json` so that this branch should
work as a proof-of-concept until those changes are reviewed and merged
by someone else. The only reminaing problem in this branch is that I'm
using a liberal date validator for CSV parsing because (I think?) the
mock data that I was given had dates in formats that were previously not
allowed.